# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResult:
    properties:
      message:
        example: Error message
        type: string
      status:
        example: Error
        type: string
    type: object
  models.GenericErrorResponse:
    properties:
      result:
        $ref: "#/definitions/models.ErrorResult"
    type: object
...
  models.Space:
    properties:
      createdAt:
        type: string
      id:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
      userId:
        type: string
    required:
      - createdAt
      - modifiedAt
      - name
    type: object
  models.SpaceRequest:
    properties:
      createdAt:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
    required:
      - createdAt
      - modifiedAt
      - name
    type: object
  models.SpaceResponseForSwagger:
    properties:
      message:
        example: Success message
        type: string
      result:
        $ref: "#/definitions/models.Space"
      status:
        example: Success
        type: string
    type: object
...
  models.Tag:
    properties:
      createdAt:
        type: string
      id:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
      userId:
        type: string
    required:
      - createdAt
      - modifiedAt
      - name
    type: object
  models.TagRequest:
    properties:
      createdAt:
        type: string
      modifiedAt:
        type: string
      name:
        type: string
    required:
      - createdAt
      - modifiedAt
      - name
    type: object
  models.TagResponseForSwagger:
    properties:
      message:
        example: Success message
        type: string
      result:
        $ref: "#/definitions/models.Tag"
      status:
        example: Success
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Blockstracker API
  title: Blockstracker
  version: "1.0"
paths:
...
  /spaces:
    post:
      consumes:
        - application/json
      description: Create a new Space with the given details
      parameters:
        - description: Space details
          in: body
          name: Space
          required: true
          schema:
            $ref: "#/definitions/models.SpaceRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.SpaceResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Create a new Space
      tags:
        - spaces
  /spaces/{id}:
    put:
      consumes:
        - application/json
      description: Update an existing Space with the given details
      parameters:
        - description: Space ID
          in: path
          name: id
          required: true
          type: string
        - description: Space details
          in: body
          name: space
          required: true
          schema:
            $ref: "#/definitions/models.SpaceRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.SpaceResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Update an existing Space
      tags:
        - spaces
  /tags:
    post:
      consumes:
        - application/json
      description: Create a new tag with the given details
      parameters:
        - description: Tag details
          in: body
          name: tag
          required: true
          schema:
            $ref: "#/definitions/models.TagRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.TagResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Create a new tag
      tags:
        - tags
  /tags/{id}:
    put:
      consumes:
        - application/json
      description: Update an existing tag with the given details
      parameters:
        - description: Tag ID
          in: path
          name: id
          required: true
          type: string
        - description: Tag details
          in: body
          name: tag
          required: true
          schema:
            $ref: "#/definitions/models.TagRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.TagResponseForSwagger"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.GenericErrorResponse"
      summary: Update an existing tag
      tags:
        - tags
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"