# models/space.go
package models

import (
	"github.com/google/uuid"
	"gorm.io/gorm"
)

type Space struct {
	ID           uuid.UUID      `gorm:"type:uuid;default:gen_random_uuid();primaryKey" json:"id"`
	Name         string         `json:"name" binding:"required"`
	CreatedAt    JSONTime       `json:"createdAt" binding:"required"`
	ModifiedAt   JSONTime       `json:"modifiedAt" binding:"required"`
	DeletedAt    gorm.DeletedAt `gorm:"index" json:"-"`
	UserID       uuid.UUID      `gorm:"type:uuid;index" json:"userId"`
	LastChangeID int64          `gorm:"not null;default:0" json:"lastChangeId"`
}

type SpaceRequest struct {
	ID         uuid.UUID `json:"id" binding:"required,uuid"`
	Name       string    `json:"name" binding:"required"`
	CreatedAt  JSONTime  `json:"createdAt" binding:"required"`
	ModifiedAt JSONTime  `json:"modifiedAt" binding:"required"`
}

type SpaceResponseForSwagger struct {
	Result Space `json:"result"`
	SuccessResult
}